+++ src/test/java/com/planbase/pdf/layoutmanager/pages/SinglePageTest.kt
@@ -1,5 +1,6 @@
 package com.planbase.pdf.layoutmanager.pages
 
+import TestManuallyPdfLayoutMgr.Companion.RGB_DARK_GRAY
 import TestManuallyPdfLayoutMgr.Companion.RGB_LIGHT_BLUE
 import TestManuallyPdfLayoutMgr.Companion.RGB_LIGHT_GREEN
 import TestManuallyPdfLayoutMgr.Companion.letterLandscapeBody
@@ -15,6 +16,7 @@ import com.planbase.pdf.layoutmanager.pages.PageGroupingTest.Companion.melonHeig
 import com.planbase.pdf.layoutmanager.pages.PageGroupingTest.Companion.melonWidth
 import com.planbase.pdf.layoutmanager.utils.Coord
 import com.planbase.pdf.layoutmanager.utils.Dim
+import com.planbase.pdf.layoutmanager.utils.LineJoinStyle
 import com.planbase.pdf.layoutmanager.utils.RGB_BLACK
 import org.apache.pdfbox.cos.COSString
 import org.apache.pdfbox.pdmodel.common.PDRectangle.LETTER
@@ -33,7 +35,7 @@ class SinglePageTest {
     @Test fun testBasics() {
         val pageMgr = PdfLayoutMgr(PDDeviceRGB.INSTANCE, Dim(LETTER))
         val lp = pageMgr.startPageGrouping(LANDSCAPE, letterLandscapeBody)
-        val page:SinglePage = pageMgr.page(0)
+        val pg:SinglePage = pageMgr.page(0)
 
         val squareDim = Dim(squareSide, squareSide)
 
@@ -46,21 +48,21 @@ class SinglePageTest {
         var y = lp.yBodyTop() - melonHeight
 
         while (y >= lp.yBodyBottom) {
-            val imgHaP:HeightAndPage = page.drawImage(Coord(melonX, y), bigMelon)
+            val imgHaP: HeightAndPage = pg.drawImage(Coord(melonX, y), bigMelon)
             assertEquals(melonHeight, imgHaP.height, 0.0)
 
-            val txtHaP:HeightAndPage = page.drawStyledText(Coord(textX, y), bigText.text, bigText.textStyle, true)
+            val txtHaP: HeightAndPage = pg.drawStyledText(Coord(textX, y), bigText.text, bigText.textStyle, true)
             assertEquals(bigText.textStyle.lineHeight, txtHaP.height, 0.0)
 
-            val rectY = page.fillRect(Coord(squareX, y), squareDim, RGB_BLACK, true)
+            val rectY: Double = pg.fillRect(Coord(squareX, y), squareDim, RGB_BLACK, true)
             assertEquals(squareSide, rectY, 0.0)
 
-            diamondRect(page, Coord(lineX1, y), squareSide)
+            diamondRect(pg, Coord(lineX1, y), squareSide)
 
-            val cellDaP: DimAndPageNums = qbfCell.render(page, Coord(cellX1, y + qbfCell.dim.height))
+            val cellDaP: DimAndPageNums = qbfCell.render(pg, Coord(cellX1, y + qbfCell.dim.height))
             Dim.assertEquals(qbfCell.dim, cellDaP.dim, 0.0)
 
-            val tableDaP: DimAndPageNums = qbfTable.render(page, Coord(tableX1, y + qbfCell.dim.height))
+            val tableDaP: DimAndPageNums = qbfTable.render(pg, Coord(tableX1, y + qbfCell.dim.height))
             Dim.assertEquals(qbfTable.dim, tableDaP.dim, 0.0)
 
             y -= melonHeight
@@ -127,13 +129,15 @@ fun diamondRect(page:RenderTarget, lowerLeft: Coord, size:Double) {
                               Coord(xMid, yBot),
                               Coord(xLeft, yMid),
                               midTop),
-                       ls)
+                       ls, LineJoinStyle.BEVEL)
 }
 
 const val squareSide = 70.0
 val times15 = TextStyle(PDType1Font.TIMES_ROMAN, 15.0, RGB_BLACK)
+val thickLine = LineStyle(RGB_DARK_GRAY, 4.0)
 val paleGreenLeft = CellStyle(Align.TOP_LEFT,
-                              BoxStyle(Padding(2.0), RGB_LIGHT_GREEN, BorderStyle(RGB_BLACK)))
+                              BoxStyle(Padding(1.0), RGB_LIGHT_GREEN,
+                                       BorderStyle(thickLine,thickLine,thickLine,thickLine,LineJoinStyle.ROUND)))
 val paleBlueLeft = CellStyle(Align.TOP_LEFT,
                              BoxStyle(Padding(2.0), RGB_LIGHT_BLUE, BorderStyle(RGB_BLACK)))
 val qbfCell = Cell(paleGreenLeft, squareSide,