+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Bigger headings mean more stable releases!
    This gives access to the functionality described in PDF 32000-1:2008 section 8.4.3.4: "Line Join Style."
  - RenderTarget.drawLine(), .drawLineStrip(), and .drawLineLoop() now take a lineJoinStyle parameter.
  - RenderTarget.drawLineLoop() now also takes a nullable backgroundColor parameter.  It will fill the shape with the background color if supplied.
+ - NOTE: For now, this only works for cells or lines that don't cross page breaks.
+ Cells or lines drawn across page boundaries will revert to either MITERed or no join style (and no fill style).
  - RenderTarget has helper functions so that you don't need to supply the new parameters if you don't want to.
  If you call the old functions with a final parameter reallyRender=true you can just remove that parameter
  to use the new helper functions.  Otherwise, you'll have to pass the new parameters as well.